+++ mw/base.ts
@@ -31,8 +31,8 @@ declare interface IWikitextModel extends editor.ITextModel {
 }
 
 // 每次新增插件都需要修改这里
-const baseVersion = '2.21',
-	addons = ['hover', 'signatureHelp'];
+const baseVersion = '2.22',
+	addons = ['inlayHints'];
 
 mw.loader.load(`${CDN}/${REPO_CDN}/mediawiki.css`, 'text/css');
 
@@ -84,6 +84,7 @@ const linters: Record<string, LintSource | undefined> = {},
 		['scrollPastEnd', 'scrollBeyondLastLine', false, true],
 		['hover', 'hover', {enabled: false}, undefined],
 		['signatureHelp', 'parameterHints', {enabled: false}, undefined],
+		['inlayHints', 'inlayHints', {enabled: 'offUnlessPressed'}, {enabled: 'onUnlessPressed'}],
 	];
 
 /**